+++ packages/frontend/src/features/sqlRunner/components/Sidebar.tsx
@@ -1,7 +1,13 @@
-import { ActionIcon, Group, Stack, Title, Tooltip } from '@mantine/core';
+import {
+    ActionIcon,
+    Group,
+    ScrollArea,
+    Stack,
+    Title,
+    Tooltip,
+} from '@mantine/core';
 import { IconLayoutSidebarLeftCollapse } from '@tabler/icons-react';
 import { type Dispatch, type FC, type SetStateAction } from 'react';
-import { ConditionalVisibility } from '../../../components/common/ConditionalVisibility';
 import MantineIcon from '../../../components/common/MantineIcon';
 import { useAppSelector } from '../store/hooks';
 import { SidebarTabs } from '../store/sqlRunnerSlice';
@@ -35,21 +41,31 @@ export const Sidebar: FC<Props> = ({ setSidebarOpen }) => {
                 </Tooltip>
             </Group>
 
-            <ConditionalVisibility
-                isVisible={activeSidebarTab === SidebarTabs.TABLES}
+            <Stack
+                display={
+                    activeSidebarTab === SidebarTabs.TABLES ? 'inherit' : 'none'
+                }
+                sx={{ flex: 1, overflow: 'hidden' }}
             >
-                <Stack sx={{ flex: 1, overflow: 'hidden' }}>
-                    <TablesPanel />
-                </Stack>
-            </ConditionalVisibility>
+                <TablesPanel />
+            </Stack>
 
-            <ConditionalVisibility
-                isVisible={activeSidebarTab === SidebarTabs.VISUALIZATION}
+            <ScrollArea
+                offsetScrollbars
+                variant="primary"
+                className="only-vertical"
+                sx={{
+                    flex: 1,
+                    display:
+                        activeSidebarTab === SidebarTabs.VISUALIZATION
+                            ? 'inherit'
+                            : 'none',
+                }}
             >
                 <Stack sx={{ flex: 1, overflow: 'hidden' }}>
                     <VisualizationConfigPanel />
                 </Stack>
-            </ConditionalVisibility>
+            </ScrollArea>
         </Stack>
     );
 };